requirements.txt
#===================================================================
# This file is auto generated by PackageIt
# Be careful to manually add modules as it might be over written.
# Check the [Project Import][ReWrite] settings in:
# packageit\pimp.ini and
# D:\Dropbox\Projects\GitIt\.packageit\packageit.ini
#===================================================================
beetools
gitpython
pydantic
